requirements.txt
streamlit
opencv-python-headless
easyocr
deep-translator
pillow
numpy
